Moved the dashboard folder out of the user
