+++ src/app/app-routing.module.ts
@@ -5,7 +5,7 @@ import { LoginComponent } from './authorisation/login/login.component';
 import { NewUserComponent } from './authorisation/new-user/new-user.component';
 import { HomeComponent } from './home/home.component';
 import { UserComponent } from './user/user.component';
-import { DasboardComponent } from './user/dasboard/dasboard.component';
+import { DasboardComponent } from './dasboard/dasboard.component';
 import { AngularFireAuthGuard, redirectLoggedInTo, redirectUnauthorizedTo } from '@angular/fire/compat/auth-guard';
 import { ProfileComponent } from './user/profile/profile.component';
